cargo clippy --fix unnecessary_cast

+++ rust-doomgeneric/engine/src/d_net.rs
@@ -34,7 +34,7 @@ unsafe fn PlayerQuitGame(state: &mut GameState, mut player: *mut player_t) {
 }
 unsafe fn RunTic(state: &mut GameState, mut cmds: *mut ticcmd_t, mut ingame: *mut bool) {
     let mut i: u32 = 0;
-    i = 0 as u32;
+    i = 0_u32;
     while i < MAXPLAYERS as u32 {
         if !state.g_game.demoplayback
             && state.g_game.playeringame[i as usize]
@@ -74,7 +74,7 @@ unsafe fn LoadGameSettings(state: &mut GameState, mut settings: *mut net_gameset
             "NOTE: Turning resolution is reduced; this is probably because there is a client recording a Vanilla demo."
         );
     }
-    i = 0 as u32;
+    i = 0_u32;
     while i < MAXPLAYERS as u32 {
         state.g_game.playeringame[i as usize] = i < (*settings).num_players as u32;
         i = i.wrapping_add(1);
@@ -91,26 +91,26 @@ unsafe fn SaveGameSettings(state: &mut GameState, mut settings: *mut net_gameset
     (*settings).fast_monsters = state.d_main.fastparm as i32;
     (*settings).respawn_monsters = state.d_main.respawnparm as i32;
     (*settings).timelimit = state.g_game.timelimit;
-    (*settings).lowres_turn = (M_CheckParm(state, "-record") > 0 as i32
-        && M_CheckParm(state, "-longtics") == 0 as i32) as i32;
+    (*settings).lowres_turn = (M_CheckParm(state, "-record") > 0_i32
+        && M_CheckParm(state, "-longtics") == 0_i32) as i32;
 }
 unsafe fn InitConnectData(state: &mut GameState, mut connect_data: *mut net_connect_data_t) {
     (*connect_data).max_players = MAXPLAYERS;
     (*connect_data).drone = false_0;
-    if M_CheckParm(state, "-left") > 0 as i32 {
+    if M_CheckParm(state, "-left") > 0_i32 {
         state.r_main.viewangleoffset = ANG90;
         (*connect_data).drone = true_0;
     }
-    if M_CheckParm(state, "-right") > 0 as i32 {
+    if M_CheckParm(state, "-right") > 0_i32 {
         state.r_main.viewangleoffset = ANG270 as i32;
         (*connect_data).drone = true_0;
     }
     (*connect_data).gamemode = state.doomstat.gamemode as i32;
     (*connect_data).gamemission = state.doomstat.gamemission as i32;
-    (*connect_data).lowres_turn = (M_CheckParm(state, "-record") > 0 as i32
-        && M_CheckParm(state, "-longtics") == 0 as i32) as i32;
+    (*connect_data).lowres_turn = (M_CheckParm(state, "-record") > 0_i32
+        && M_CheckParm(state, "-longtics") == 0_i32) as i32;
     W_Checksum(state, &raw mut (*connect_data).wad_sha1sum as *mut byte);
-    (*connect_data).is_freedoom = (W_CheckNumForName(&mut state.w_wad, "FREEDOOM") >= 0 as i32) as i32;
+    (*connect_data).is_freedoom = (W_CheckNumForName(&mut state.w_wad, "FREEDOOM") >= 0_i32) as i32;
 }
 pub unsafe fn D_ConnectNetGame(state: &mut GameState) {
     let mut connect_data: net_connect_data_t = net_connect_data_t {
@@ -126,7 +126,7 @@ pub unsafe fn D_ConnectNetGame(state: &mut GameState) {
     };
     InitConnectData(state, &raw mut connect_data);
     state.g_game.netgame = D_InitNetGame(state, &raw mut connect_data);
-    if M_CheckParm(state, "-solo-net") > 0 as i32 {
+    if M_CheckParm(state, "-solo-net") > 0_i32 {
         state.g_game.netgame = true;
     }
 }
@@ -167,16 +167,16 @@ pub unsafe fn D_CheckNetGame(state: &mut GameState) {
     );
     println!(
         "player {} of {} ({} nodes)",
-        state.g_game.consoleplayer + 1 as i32,
+        state.g_game.consoleplayer + 1_i32,
         settings.num_players,
         settings.num_players,
     );
-    if state.g_game.timelimit > 0 as i32 && state.g_game.deathmatch != 0 {
-        if state.g_game.timelimit == 20 as i32 && M_CheckParm(state, "-avg") != 0 {
+    if state.g_game.timelimit > 0_i32 && state.g_game.deathmatch != 0 {
+        if state.g_game.timelimit == 20_i32 && M_CheckParm(state, "-avg") != 0 {
             println!("Austin Virtual Gaming: Levels will end after 20 minutes");
         } else {
             print!("Levels will end after {} minute", state.g_game.timelimit);
-            if state.g_game.timelimit > 1 as i32 {
+            if state.g_game.timelimit > 1_i32 {
                 print!("s");
             }
             println!(".");
