+++ rust-doomgeneric/engine/src/i_cdmus.rs
@@ -9,27 +9,27 @@ impl ICdMusState {
 }
 
 pub fn I_CDMusInit() -> i32 {
-    return 0 as i32;
+    return 0_i32;
 }
 pub fn I_CDMusPlay() -> i32 {
-    return 0 as i32;
+    return 0_i32;
 }
 pub fn I_CDMusStop() -> i32 {
-    return 0 as i32;
+    return 0_i32;
 }
 pub fn I_CDMusResume() -> i32 {
-    return 0 as i32;
+    return 0_i32;
 }
 pub fn I_CDMusSetVolume(state: &mut ICdMusState) -> i32 {
-    state.cd_Error = 0 as i32;
-    return 0 as i32;
+    state.cd_Error = 0_i32;
+    return 0_i32;
 }
 pub fn I_CDMusFirstTrack() -> i32 {
-    return 0 as i32;
+    return 0_i32;
 }
 pub fn I_CDMusLastTrack() -> i32 {
-    return 0 as i32;
+    return 0_i32;
 }
 pub fn I_CDMusTrackLength() -> i32 {
-    return 0 as i32;
+    return 0_i32;
 }
