+++ rust-doomgeneric/engine/src/m_argv.rs
@@ -12,7 +12,7 @@ impl MArgvState {
 
 pub const DIR_SEPARATOR: char = '/';
 pub fn M_CheckParmWithArgs(state: &mut GameState, check: &str, mut num_args: i32) -> i32 {
-    let mut i: i32 = 1 as i32;
+    let mut i: i32 = 1_i32;
     while i < state.m_argv.myargv.len() as i32 - num_args {
         if state.m_argv.myargv[i as usize]
             .to_str()
@@ -22,16 +22,16 @@ pub fn M_CheckParmWithArgs(state: &mut GameState, check: &str, mut num_args: i32
         }
         i += 1;
     }
-    return 0 as i32;
+    return 0_i32;
 }
 pub fn M_ParmExists(state: &mut GameState, check: &str) -> bool {
-    return M_CheckParm(state, check) != 0 as i32;
+    return M_CheckParm(state, check) != 0_i32;
 }
 pub fn M_CheckParm(state: &mut GameState, check: &str) -> i32 {
-    return M_CheckParmWithArgs(state, check, 0 as i32);
+    return M_CheckParmWithArgs(state, check, 0_i32);
 }
 pub fn M_FindResponseFile(state: &mut GameState) {
-    let mut i: i32 = 1 as i32;
+    let mut i: i32 = 1_i32;
     while i < state.m_argv.myargv.len() as i32 {
         i += 1;
     }
