+++ rust-doomgeneric/engine/src/m_fixed.rs
@@ -3,7 +3,7 @@ pub const INT_MAX: i32 = i32::MAX;
 pub const INT_MIN: i32 = i32::MIN;
 pub const FRACBITS: i32 = 16;
 pub fn FixedMul(a: fixed_t, b: fixed_t) -> fixed_t {
-    (a as i64 * b as i64 >> FRACBITS) as fixed_t
+    ((a as i64 * b as i64) >> FRACBITS) as fixed_t
 }
 pub fn FixedDiv(a: fixed_t, b: fixed_t) -> fixed_t {
     if a.abs() >> 14 >= b.abs() {
@@ -13,7 +13,7 @@ pub fn FixedDiv(a: fixed_t, b: fixed_t) -> fixed_t {
             INT_MAX
         }
     } else {
-        let result: i64 = ((a as i64) << 16 as i32) / b as i64;
+        let result: i64 = ((a as i64) << 16_i32) / b as i64;
         result as fixed_t
     }
 }
