+++ rust-doomgeneric/engine/src/p_tick.rs
@@ -271,11 +271,11 @@ pub unsafe fn P_Ticker(state: &mut GameState) {
     if !state.g_game.netgame
         && state.m_menu.menuactive
         && !state.g_game.demoplayback
-        && state.g_game.players[state.g_game.consoleplayer as usize].viewz != 1 as i32
+        && state.g_game.players[state.g_game.consoleplayer as usize].viewz != 1_i32
     {
         return;
     }
-    i = 0 as i32;
+    i = 0_i32;
     while i < MAXPLAYERS {
         if state.g_game.playeringame[i as usize] {
             let player: *mut player_t = &mut state.g_game.players[i as usize];
